+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## V 1.5.0 - XX/XX/19
 
+    - ControlComponent : Add MOUSEFOLLOW ControlType
     - Entry : Add Ctrl-C and Ctrl-V
 
 ## V 1.4.2 - 17/07/19